+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## unreleased
 * Deprecate `deviceSessionId` and `fraudMerchantId` in `CreditCardGateway`, `CustomerGateway`, `PaymentMethodGateway`, and `TransactionGateway` classes
+* Add `installments` to Transaction sale
+* Add `count` to `installments`
 
 ## 5.4.0
 * Add `AcquirerReferenceNumber` to `Transaction`